HS Project copy/requirements.txt
asgiref==3.3.2
astroid==2.4.2
botocore==1.19.43
cement==2.8.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-allauth==0.43.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
future==0.16.0
idna==2.10
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.5.9
Pillow==8.0.1
# psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
# pypiwin32==223
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.1
# pywin32==300
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.2
urllib3==1.25.11
wcwidth==0.1.9
wrapt==1.12.1